requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.1
django-appconf==1.1.0
django-autoslug==1.9.9
mysqlclient==2.2.7
pilkit==3.0
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
